+++ CHANGELOG.md
@@ -7,6 +7,12 @@
 ### Fixed
 -`<StackedAreaChart />` stroke is cut off if it reaches the top of the chart
 
+### Fixed
+- Various xAxis label bugs that caused labels to truncate too early
+
+### Added 
+- Charts with `xAxisOptions` have the option to disable text wrapping of xAxis labels
+
 ## [0.21.3] - 2021-10-15
 
 ### Fixed